Environment to run timetagger programs with visual studio and python plotting program
xtdc_babel contains the necessary files to run a program for the time tagger. The .cpp file must be in src in the timtagger_example folder and the executable name and file name of the program must be modified in the CMakelists that is in the tools folder. The executable will appear in lib/x64/Debug.