Add files via upload
This drivers need to be here for the time tagger executable to work. By building in visual studio from the timetagger example folder we can now get an executable in this folder. It will not immediatly work with the python gui because the output name when building will now be get_data3. This is to prevent erasing the functioning executable file.