Working TimeTagger code integrated with SFML histogram
The cmake lists is now able to compile the cronologic libraries alongside the SFML library for windows, the histogram is now integrated with the Time Tagger code, which allows for data to be instantly plotted as it comes in. The loop can be run continuously and at the end the data acquired from  the session is stored in a .txt file for future analysis.